backend/requirements.txt
# Backend Dependencies - Massive API Version - Pls keep updated - Arlor09
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0
python-dotenv==1.0.0